Add ModalStore Tests
