+++ src/store/ModalStore.js
@@ -1,19 +1,4 @@
 import { types } from 'mobx-state-tree'
-import UnapproveModal from 'components/Modals/UnapproveModal'
-import DownloadData from 'components/Modals/DownloadDataModal'
-import SubjectLocked from 'components/Modals/SubjectLockedModal'
-import SearchModal from 'components/SearchModal'
-
-const modalTypes = {
-  UNAPPROVED: UnapproveModal,
-  DOWNLOAD: DownloadData,
-  LOCKED: SubjectLocked,
-  SEARCH: SearchModal
-}
-
-function getModal(type) {
-  return modalTypes[type] || null
-}
 
 const ModalStore = types.model('ModalStore', {
   current: types.optional(types.string, ''),
@@ -23,4 +8,4 @@ const ModalStore = types.model('ModalStore', {
   }
 }))
 
-export { getModal, ModalStore }
+export { ModalStore }
